# p2p, refactor: performance improvements to ProtectEvictionCandidatesByRatio()