This follow-up to #21261 improves `ProtectEvictionCandidatesByRatio()` for better performance. 

Benchmarks are added; the performance improvement is between 2x and 5x for the benchmarked cases (CPU 2.50GHz, Turbo off, performance mode, Debian Clang 11 non-debug build).

```
$ ./src/bench/bench_bitcoin -filter="EvictionProtection*.*"
```

The refactored code is well-covered by existing unit tests and also a fuzzer.

- `$ ./src/test/test_bitcoin -t net_peer_eviction_tests`
- `$ FUZZ=node_eviction ./src/test/fuzz/fuzz ../qa-assets/fuzz_seed_corpus/node_eviction`