@jonatack 2021-06-19
`./src/test/test_bitcoin -t net_peer_eviction_tests/peer_protection_test` runs about 10% faster for me on this branch, though it is a debug build and the unit tests use very low numbers of eviction candidates (between 4 and 24).  I'll write a test with higher numbers of eviction candidates and benchmark with optimized builds (and fix the fuzzer CI failure).