@jonatack 2021-06-19
<details><summary>With this 2-line test change</summary><p>

```diff
diff --git a/src/test/net_peer_eviction_tests.cpp b/src/test/net_peer_eviction_tests.cpp
index 4bfd487b86..f6ea4ba1af 100644
--- a/src/test/net_peer_eviction_tests.cpp
+++ b/src/test/net_peer_eviction_tests.cpp
@@ -85,6 +85,7 @@ BOOST_AUTO_TEST_CASE(peer_protection_test)
 {
     FastRandomContext random_context{true};
     int num_peers{12};
+    for (int i = 0; i < 500000; ++i) {
 
     // Expect half of the peers with greatest uptime (the lowest nTimeConnected)
     // to be protected from eviction.
@@ -456,6 +457,7 @@ BOOST_AUTO_TEST_CASE(peer_protection_test)
         /* protected_peer_ids */ {0, 1, 2, 3, 4, 5, 7, 8, 11, 12, 16, 17},
         /* unprotected_peer_ids */ {6, 9, 10, 13, 14, 15, 18, 19, 20, 21, 22, 23},
         random_context));
+    }
 }
```
</p></details>

```
$ time ./src/test/test_bitcoin -t net_peer_eviction_tests/peer_protection_test
```

![Screenshot from 2021-06-19 23-16-38](https://user-images.githubusercontent.com/2415484/122655715-1a0d6980-d144-11eb-80c7-28cc1646369e.jpg)

Seeing roughly an 8% speedup, Clang 11 non-debug build on Debian testing, run times in seconds on this branch vs on master. So pretty much the same as the comment above.